simplify hydra; load_random doesn't need index; refactorings
