+++ configs/mapgen.yaml
@@ -1,7 +1,6 @@
 defaults:
   - mettagrid
   - mapgen: show
-  - s3: softmax
   - game/map_builder: ???
   - _self_
