--- configs/s3/softmax.yaml
@@ -1,1 +1,0 @@
-maps_root: s3://softmax-public/maps
