+++ mettagrid/map/load_random.py
@@ -1,31 +1,23 @@
-# Root map generator, based on nodes.
-
 import random
 
 from mettagrid.map.load import Load
-from mettagrid.map.utils.serialization import load_from_uri
+from mettagrid.map.utils import s3utils
 
 from .scene import SceneCfg
 
 
 class LoadRandom(Load):
     """
-    Load a random map from a list of pregenerated maps.
+    Load a random map from S3 directory.
 
-    The directory must contain a file `index.txt` that lists all the maps in the
-    directory. (Listing S3 objects would be too slow because of pagination.)
-
-    The index file can be produced with the following command:
-        python -m tools.index_s3_maps index_s3_maps.dir=s3://...
+    See also: `LoadRandomFromIndex` for a version that loads a random map from a pre-generated index.
     """
 
-    def __init__(self, index_uri: str, extra_root: SceneCfg | None = None):
-        self._index_uri = index_uri
+    def __init__(self, dir: str, extra_root: SceneCfg | None = None):
+        self._dir = dir
 
-        # For 10k maps in a directory we'd have to fetch 100Kb of index data.
-        # (Can we optimize this further by caching?)
-        index = load_from_uri(self._index_uri)
-        index = index.split("\n")
-        random_map_uri = random.choice(index)
+        uris = s3utils.list_objects(self._dir)
+        uris = [uri for uri in uris if uri.endswith(".yaml")]
+        random_map_uri = random.choice(uris)
 
         super().__init__(random_map_uri, extra_root)
